Show links to common templates
