+++ frontend/packages/kubevirt-plugin/locales/en/kubevirt-plugin.json
@@ -676,7 +676,6 @@
   "config map / secret / service account": "config map / secret / service account",
   "Serial Number": "Serial Number",
   "Add Config Map, Secret or Service Account": "Add Config Map, Secret or Service Account",
-  "View Pending Changes": "View Pending Changes",
   "Pod": "Pod",
   "No Toleration rules": "No Toleration rules",
   "No Affinity rules": "No Affinity rules",
@@ -690,6 +689,7 @@
   "No virtual machines found": "No virtual machines found",
   "See the <2>templates tab</2> to quickly create a virtual machine from the available templates.": "See the <2>templates tab</2> to quickly create a virtual machine from the available templates.",
   "Learn how to use virtual machines": "Learn how to use virtual machines",
+  "View Pending Changes": "View Pending Changes",
   "No additional support": "No additional support",
   "Support by template provider": "Support by template provider",
   "Tiny": "Tiny",
